divide FilteredPageOutput class / use root property
